+++ datacrawler/acceptance_tests/budget_finances_médico_social_test.py
@@ -4,7 +4,8 @@ from freezegun import freeze_time
 
 import pandas as pd
 import pytest
-from numpy import NaN
+from numpy import nan
+from sqlalchemy import text
 
 import datacrawler
 from datacrawler.ajoute_le_bloc_budget_et_finances_des_établissements_médico_sociaux import ajoute_le_bloc_budget_et_finances_des_établissements_médico_sociaux
@@ -70,7 +71,7 @@ class TestAjouteLeBudgetEtFinancesDesÉtablissementsMédicoSociaux:
                     numéro_finess_établissement_ca_pa,
                     numéro_finess_établissement_ca_pa,
                 ],
-                "contribution_frais_de_siege_groupement": [-22312.0, 0.0, 0.0, 0.0, NaN, NaN, NaN, NaN],
+                "contribution_frais_de_siege_groupement": [-22312.0, 0.0, 0.0, 0.0, nan, nan, nan, nan],
                 "depenses_groupe_i": [
                     -105389.53,
                     -161786,
@@ -78,8 +79,8 @@ class TestAjouteLeBudgetEtFinancesDesÉtablissementsMédicoSociaux:
                     -85102.530010000002,
                     -16062.690000000001,
                     -16114.09,
-                    NaN,
-                    NaN,
+                    nan,
+                    nan,
                 ],
                 "depenses_groupe_ii": [
                     -506251.12999999995,
@@ -88,8 +89,8 @@ class TestAjouteLeBudgetEtFinancesDesÉtablissementsMédicoSociaux:
                     -442475.08000000007,
                     -482402.46000000008,
                     -522700.76999999996,
-                    NaN,
-                    NaN,
+                    nan,
+                    nan,
                 ],
                 "depenses_groupe_iii": [
                     -88214.989999999991,
@@ -98,8 +99,8 @@ class TestAjouteLeBudgetEtFinancesDesÉtablissementsMédicoSociaux:
                     -9134.2200000000012,
                     -44491.319999999992,
                     -44619.190000000002,
-                    NaN,
-                    NaN,
+                    nan,
+                    nan,
                 ],
                 "recettes_groupe_i": [
                     628872.06999999995,
@@ -108,8 +109,8 @@ class TestAjouteLeBudgetEtFinancesDesÉtablissementsMédicoSociaux:
                     543015.84999999998,
                     588568.68999999994,
                     583376.64000000001,
-                    NaN,
-                    NaN,
+                    nan,
+                    nan,
                 ],
                 "recettes_groupe_ii": [
                     46843.48,
@@ -118,10 +119,10 @@ class TestAjouteLeBudgetEtFinancesDesÉtablissementsMédicoSociaux:
                     9410.4599999999991,
                     782.12,
                     511.08999999999997,
-                    NaN,
-                    NaN,
+                    nan,
+                    nan,
                 ],
-                "recettes_groupe_iii": [27174.48, 0, 12830, 12830, 26733.739999999998, 24276.970000000001, NaN, NaN],
+                "recettes_groupe_iii": [27174.48, 0, 12830, 12830, 26733.739999999998, 24276.970000000001, nan, nan],
                 "resultat_net_comptable": [
                     3034.3799999998928,
                     7289.9200000003912,
@@ -140,8 +141,8 @@ class TestAjouteLeBudgetEtFinancesDesÉtablissementsMédicoSociaux:
                     0.12519374136642053,
                     0.11776359918113584,
                     0.049315762194766362,
-                    NaN,
-                    NaN,
+                    nan,
+                    nan,
                 ],
                 "taux_de_vetuste_construction": [
                     0.38845089702004892,
@@ -149,22 +150,22 @@ class TestAjouteLeBudgetEtFinancesDesÉtablissementsMédicoSociaux:
                     0.56203591359317973,
                     0.56203591359317973,
                     0.51376936316695354,
-                    NaN,
+                    nan,
                     0.31154835988672847,
-                    NaN,
+                    nan,
                 ],
                 "fonds_de_roulement": [
                     2206969.259999999800000000,
                     1057217.929999999900000000,
                     3988284.410000000100000000,
                     3988284.410000000100000000,
-                    NaN,
-                    NaN,
-                    NaN,
-                    NaN,
+                    nan,
+                    nan,
+                    nan,
+                    nan,
                 ],
-                "charges": [NaN, NaN, NaN, NaN, NaN, NaN, -177631.38999999998, -207285.97000000003],
-                "produits": [NaN, NaN, NaN, NaN, NaN, NaN, 196518.51999999999, 219272.62],
+                "charges": [nan, nan, nan, nan, nan, nan, -177631.38999999998, -207285.97000000003],
+                "produits": [nan, nan, nan, nan, nan, nan, 196518.51999999999, 219272.62],
             },
         )
 
@@ -201,18 +202,18 @@ class TestAjouteLeBudgetEtFinancesDesÉtablissementsMédicoSociaux:
         )
 
         # THEN
-        date_du_fichier_ann_errd_ej_et = base_de_données_test.execute(
-            f"SELECT * FROM {TABLE_DES_MISES_À_JOUR_DES_FICHIERS_SOURCES} WHERE fichier = '{FichierSource.DIAMANT_ANN_ERRD_EJ_ET.value}'"
+        date_du_fichier_ann_errd_ej_et = base_de_données_test.connect().execute(
+            text(f"SELECT * FROM {TABLE_DES_MISES_À_JOUR_DES_FICHIERS_SOURCES} WHERE fichier = '{FichierSource.DIAMANT_ANN_ERRD_EJ_ET.value}'")
         )
         assert date_du_fichier_ann_errd_ej_et.fetchone() == (date(2022, 6, 7), FichierSource.DIAMANT_ANN_ERRD_EJ_ET.value)
 
-        date_du_fichier_ann_ca_ej_et = base_de_données_test.execute(
-            f"SELECT * FROM {TABLE_DES_MISES_À_JOUR_DES_FICHIERS_SOURCES} WHERE fichier = '{FichierSource.DIAMANT_ANN_CA_EJ_ET.value}'"
+        date_du_fichier_ann_ca_ej_et = base_de_données_test.connect().execute(
+            text(f"SELECT * FROM {TABLE_DES_MISES_À_JOUR_DES_FICHIERS_SOURCES} WHERE fichier = '{FichierSource.DIAMANT_ANN_CA_EJ_ET.value}'")
         )
         assert date_du_fichier_ann_ca_ej_et.fetchone() == (date(2022, 9, 1), FichierSource.DIAMANT_ANN_CA_EJ_ET.value)
 
-        date_du_fichier_ann_errd_ej = base_de_données_test.execute(
-            f"SELECT * FROM {TABLE_DES_MISES_À_JOUR_DES_FICHIERS_SOURCES} WHERE fichier = '{FichierSource.DIAMANT_ANN_ERRD_EJ.value}'"
+        date_du_fichier_ann_errd_ej = base_de_données_test.connect().execute(
+            text(f"SELECT * FROM {TABLE_DES_MISES_À_JOUR_DES_FICHIERS_SOURCES} WHERE fichier = '{FichierSource.DIAMANT_ANN_ERRD_EJ.value}'")
         )
         assert date_du_fichier_ann_errd_ej.fetchone() == (date(2022, 9, 1), FichierSource.DIAMANT_ANN_ERRD_EJ.value)
 
@@ -265,7 +266,7 @@ class TestAjouteLeBudgetEtFinancesDesÉtablissementsMédicoSociaux:
                     numéro_finess_établissement_ca_pa,
                     numéro_finess_établissement_ca_pa,
                 ],
-                "contribution_frais_de_siege_groupement": [-22312.0, 0.0, 0.0, 0.0, NaN, NaN, NaN, NaN],
+                "contribution_frais_de_siege_groupement": [-22312.0, 0.0, 0.0, 0.0, nan, nan, nan, nan],
                 "depenses_groupe_i": [
                     -105389.53,
                     -161786,
@@ -273,8 +274,8 @@ class TestAjouteLeBudgetEtFinancesDesÉtablissementsMédicoSociaux:
                     -85102.530010000002,
                     -16062.690000000001,
                     -16114.09,
-                    NaN,
-                    NaN,
+                    nan,
+                    nan,
                 ],
                 "depenses_groupe_ii": [
                     -506251.12999999995,
@@ -283,8 +284,8 @@ class TestAjouteLeBudgetEtFinancesDesÉtablissementsMédicoSociaux:
                     -442475.08000000007,
                     -482402.46000000008,
                     -522700.76999999996,
-                    NaN,
-                    NaN,
+                    nan,
+                    nan,
                 ],
                 "depenses_groupe_iii": [
                     -88214.989999999991,
@@ -293,8 +294,8 @@ class TestAjouteLeBudgetEtFinancesDesÉtablissementsMédicoSociaux:
                     -9134.2200000000012,
                     -44491.319999999992,
                     -44619.190000000002,
-                    NaN,
-                    NaN,
+                    nan,
+                    nan,
                 ],
                 "recettes_groupe_i": [
                     628872.06999999995,
@@ -303,8 +304,8 @@ class TestAjouteLeBudgetEtFinancesDesÉtablissementsMédicoSociaux:
                     543015.84999999998,
                     588568.68999999994,
                     583376.64000000001,
-                    NaN,
-                    NaN,
+                    nan,
+                    nan,
                 ],
                 "recettes_groupe_ii": [
                     46843.48,
@@ -313,10 +314,10 @@ class TestAjouteLeBudgetEtFinancesDesÉtablissementsMédicoSociaux:
                     9410.4599999999991,
                     782.12,
                     511.08999999999997,
-                    NaN,
-                    NaN,
+                    nan,
+                    nan,
                 ],
-                "recettes_groupe_iii": [27174.48, 0, 12830, 12830, 26733.739999999998, 24276.970000000001, NaN, NaN],
+                "recettes_groupe_iii": [27174.48, 0, 12830, 12830, 26733.739999999998, 24276.970000000001, nan, nan],
                 "resultat_net_comptable": [
                     3034.3799999998928,
                     7289.9200000003912,
@@ -335,8 +336,8 @@ class TestAjouteLeBudgetEtFinancesDesÉtablissementsMédicoSociaux:
                     0.12519374136642053,
                     0.11776359918113584,
                     0.049315762194766362,
-                    NaN,
-                    NaN,
+                    nan,
+                    nan,
                 ],
                 "taux_de_vetuste_construction": [
                     0.38845089702004892,
@@ -344,22 +345,22 @@ class TestAjouteLeBudgetEtFinancesDesÉtablissementsMédicoSociaux:
                     0.56203591359317973,
                     0.56203591359317973,
                     0.51376936316695354,
-                    NaN,
+                    nan,
                     0.31154835988672847,
-                    NaN,
+                    nan,
                 ],
                 "fonds_de_roulement": [
                     2206969.259999999800000000,
                     1057217.929999999900000000,
                     3988284.410000000100000000,
                     3988284.410000000100000000,
-                    NaN,
-                    NaN,
-                    NaN,
-                    NaN,
+                    nan,
+                    nan,
+                    nan,
+                    nan,
                 ],
-                "charges": [NaN, NaN, NaN, NaN, NaN, NaN, -177631.38999999998, -207285.97000000003],
-                "produits": [NaN, NaN, NaN, NaN, NaN, NaN, 196518.51999999999, 219272.62],
+                "charges": [nan, nan, nan, nan, nan, nan, -177631.38999999998, -207285.97000000003],
+                "produits": [nan, nan, nan, nan, nan, nan, 196518.51999999999, 219272.62],
             },
         )
 
@@ -429,7 +430,7 @@ class TestAjouteLeBudgetEtFinancesDesÉtablissementsMédicoSociaux:
                 {
                     "numero_finess_etablissement_territorial": [numéro_finess_établissement_errd, numéro_finess_établissement_ca_ph],
                     "annee": [2020, 2020],
-                    "contribution_frais_de_siege_groupement": [-300.0, NaN],
+                    "contribution_frais_de_siege_groupement": [-300.0, nan],
                     "depenses_groupe_i": [-100.0, -100.0],
                     "depenses_groupe_ii": [-200.0, -200.0],
                     "depenses_groupe_iii": [-300.0, -300.0],
@@ -440,19 +441,19 @@ class TestAjouteLeBudgetEtFinancesDesÉtablissementsMédicoSociaux:
                     "cadre_budgetaire": ["ERRD", "CA_PH"],
                     "taux_de_caf": [0.071600138178413528, 0.16],
                     "taux_de_vetuste_construction": [0.45555983373892417, 0.53],
-                    "fonds_de_roulement": [2206969.259999999800000000, NaN],
-                    "produits": [NaN, NaN],
-                    "charges": [NaN, NaN],
+                    "fonds_de_roulement": [2206969.259999999800000000, nan],
+                    "produits": [nan, nan],
+                    "charges": [nan, nan],
                 }
             ).sort_index(axis=1),
         )
 
-        date_du_fichier_de_données_errd = base_de_données_test.execute(
-            f"""SELECT * FROM {TABLE_DES_MISES_À_JOUR_DES_FICHIERS_SOURCES} WHERE fichier = '{FichierSource.DIAMANT_ANN_ERRD_EJ_ET.value}'"""
+        date_du_fichier_de_données_errd = base_de_données_test.connect().execute(
+            text(f"""SELECT * FROM {TABLE_DES_MISES_À_JOUR_DES_FICHIERS_SOURCES} WHERE fichier = '{FichierSource.DIAMANT_ANN_ERRD_EJ_ET.value}'""")
         )
         assert date_du_fichier_de_données_errd.fetchone() == (date(2020, 1, 1), FichierSource.DIAMANT_ANN_ERRD_EJ_ET.value)
 
-        date_du_fichier_de_données_ca = base_de_données_test.execute(
-            f"""SELECT * FROM {TABLE_DES_MISES_À_JOUR_DES_FICHIERS_SOURCES} WHERE fichier = '{FichierSource.DIAMANT_ANN_CA_EJ_ET.value}'"""
+        date_du_fichier_de_données_ca = base_de_données_test.connect().execute(
+            text(f"""SELECT * FROM {TABLE_DES_MISES_À_JOUR_DES_FICHIERS_SOURCES} WHERE fichier = '{FichierSource.DIAMANT_ANN_CA_EJ_ET.value}'""")
         )
         assert date_du_fichier_de_données_ca.fetchone() == (date(2020, 2, 2), FichierSource.DIAMANT_ANN_CA_EJ_ET.value)