+++ datacrawler/acceptance_tests/ressources_humaines_médico_social_test.py
@@ -4,7 +4,8 @@ from unittest.mock import Mock, patch
 import pandas as pd
 from freezegun import freeze_time
 import pytest
-from numpy import NaN
+from numpy import nan
+from sqlalchemy import text
 
 import datacrawler
 from datacrawler.ajoute_le_bloc_ressources_humaines_des_établissements_médico_sociaux import (
@@ -65,18 +66,18 @@ class TestAjouteLeBlocDesRessourcesHumainesMédicoSocial:
                     numéro_finess_ca,
                 ],
                 "annee": [2019, 2018, 2021, 2020, 2019, 2018, 2019, 2018],
-                "nombre_cdd_remplacement": [2.0, 19.0, 5.0, 5.0, 5.0, 5.0, NaN, NaN],
-                "taux_etp_vacants": [0.1197, 0.0483, 0.0, 0.0, 0.0, 0.0, NaN, NaN],
-                "taux_prestation_externes": [0.0232, NaN, 0.0164, 0.0164, 0.0082, 0.0164, NaN, NaN],
-                "taux_rotation_personnel": [0.1923, 0.1429, 0.0352, 0.0352, 0.0141, 0.0352, NaN, NaN],
-                "taux_absenteisme_maladie_courte_duree": [0.0028, 0.0021, 0.0083, 0.0083, 0.0125, 0.0083, NaN, NaN],
-                "taux_absenteisme_maladie_moyenne_duree": [0.0465, 0.0717, 0.0166, 0.0166, 0.0149, 0.0166, NaN, NaN],
-                "taux_absenteisme_maladie_longue_duree": [0.0, .1194, 0.0089, 0.0089, 0.0319, 0.0089 , NaN, NaN],
-                "taux_absenteisme_maternite_paternite": [0.0, 0.0, 0.0128, 0.0128, 0.0005, 0.0128, NaN, NaN],
-                "taux_absenteisme_accident_maladie_professionnelle": [0.0008, 0.0246, 0.0085, 0.0085, 0.0088, 0.0085, NaN, NaN],
-                "taux_absenteisme_conges_speciaux": [0.0109, 0.0, 0.0004, 0.0004, 0.0, 0.0004, NaN, NaN],
-                "taux_absenteisme_hors_formation": [0.0609, 0.2179, 0.0554, 0.0554, 0.0685, 0.0554, NaN, NaN],
-                "nombre_etp_realises": [NaN, NaN, 188.5, 50.64999999999999, 17.19, 4.55, 9.71, 10.34],
+                "nombre_cdd_remplacement": [2.0, 19.0, 5.0, 5.0, 5.0, 5.0, nan, nan],
+                "taux_etp_vacants": [0.1197, 0.0483, 0.0, 0.0, 0.0, 0.0, nan, nan],
+                "taux_prestation_externes": [0.0232, nan, 0.0164, 0.0164, 0.0082, 0.0164, nan, nan],
+                "taux_rotation_personnel": [0.1923, 0.1429, 0.0352, 0.0352, 0.0141, 0.0352, nan, nan],
+                "taux_absenteisme_maladie_courte_duree": [0.0028, 0.0021, 0.0083, 0.0083, 0.0125, 0.0083, nan, nan],
+                "taux_absenteisme_maladie_moyenne_duree": [0.0465, 0.0717, 0.0166, 0.0166, 0.0149, 0.0166, nan, nan],
+                "taux_absenteisme_maladie_longue_duree": [0.0, 0.1194, 0.0089, 0.0089, 0.0319, 0.0089, nan, nan],
+                "taux_absenteisme_maternite_paternite": [0.0, 0.0, 0.0128, 0.0128, 0.0005, 0.0128, nan, nan],
+                "taux_absenteisme_accident_maladie_professionnelle": [0.0008, 0.0246, 0.0085, 0.0085, 0.0088, 0.0085, nan, nan],
+                "taux_absenteisme_conges_speciaux": [0.0109, 0.0, 0.0004, 0.0004, 0.0, 0.0004, nan, nan],
+                "taux_absenteisme_hors_formation": [0.0609, 0.2179, 0.0554, 0.0554, 0.0685, 0.0554, nan, nan],
+                "nombre_etp_realises": [nan, nan, 188.5, 50.64999999999999, 17.19, 4.55, 9.71, 10.34],
             },
         )
 
@@ -85,7 +86,15 @@ class TestAjouteLeBlocDesRessourcesHumainesMédicoSocial:
             base_de_données_test,
         )
 
-        pd.testing.assert_frame_equal(données_ressources_humaines_enregistrées.sort_index(axis=1), données_ressources_humaines_attendues.sort_index(axis=1))
+        # On trie les data pour ne vérifier que le contenu et non l’ordre de récupération des données
+        pd.testing.assert_frame_equal(
+            données_ressources_humaines_enregistrées.sort_index(axis=1)
+            .sort_values(by=["numero_finess_etablissement_territorial", "annee"])
+            .reset_index(drop=True),
+            données_ressources_humaines_attendues.sort_index(axis=1)
+            .sort_values(by=["numero_finess_etablissement_territorial", "annee"])
+            .reset_index(drop=True),
+        )
 
     def test_sauvegarde_les_dates_de_mises_à_jour_des_indicateurs_ressources_humaines(self) -> None:
         # GIVEN
@@ -106,18 +115,18 @@ class TestAjouteLeBlocDesRessourcesHumainesMédicoSocial:
         )
 
         # THEN
-        date_du_fichier_ann_ca_ej_et = base_de_données_test.execute(
-            f"SELECT * FROM {TABLE_DES_MISES_À_JOUR_DES_FICHIERS_SOURCES} WHERE fichier = '{FichierSource.DIAMANT_ANN_CA_EJ_ET.value}'"
+        date_du_fichier_ann_ca_ej_et = base_de_données_test.connect().execute(
+            text(f"SELECT * FROM {TABLE_DES_MISES_À_JOUR_DES_FICHIERS_SOURCES} WHERE fichier = '{FichierSource.DIAMANT_ANN_CA_EJ_ET.value}'")
         )
         assert date_du_fichier_ann_ca_ej_et.fetchone() == (date(2022, 9, 1), FichierSource.DIAMANT_ANN_CA_EJ_ET.value)
 
-        date_du_fichier_ann_errd_ej_et = base_de_données_test.execute(
-            f"SELECT * FROM {TABLE_DES_MISES_À_JOUR_DES_FICHIERS_SOURCES} WHERE fichier = '{FichierSource.DIAMANT_ANN_ERRD_EJ_ET.value}'"
+        date_du_fichier_ann_errd_ej_et = base_de_données_test.connect().execute(
+            text(f"SELECT * FROM {TABLE_DES_MISES_À_JOUR_DES_FICHIERS_SOURCES} WHERE fichier = '{FichierSource.DIAMANT_ANN_ERRD_EJ_ET.value}'")
         )
         assert date_du_fichier_ann_errd_ej_et.fetchone() == (date(2022, 6, 7), FichierSource.DIAMANT_ANN_ERRD_EJ_ET.value)
 
-        date_du_fichier_ann_ms_tdp_et = base_de_données_test.execute(
-            f"SELECT * FROM {TABLE_DES_MISES_À_JOUR_DES_FICHIERS_SOURCES} WHERE fichier = '{FichierSource.DIAMANT_ANN_MS_TDP_ET.value}'"
+        date_du_fichier_ann_ms_tdp_et = base_de_données_test.connect().execute(
+            text(f"SELECT * FROM {TABLE_DES_MISES_À_JOUR_DES_FICHIERS_SOURCES} WHERE fichier = '{FichierSource.DIAMANT_ANN_MS_TDP_ET.value}'")
         )
         assert date_du_fichier_ann_ms_tdp_et.fetchone() == (date(2022, 6, 7), FichierSource.DIAMANT_ANN_MS_TDP_ET.value)
 
@@ -163,21 +172,21 @@ class TestAjouteLeBlocDesRessourcesHumainesMédicoSocial:
                     NUMÉRO_FINESS_ÉTABLISSEMENT_MÉDICO_SOCIAL,
                     NUMÉRO_FINESS_ÉTABLISSEMENT_MÉDICO_SOCIAL,
                     NUMÉRO_FINESS_ÉTABLISSEMENT_MÉDICO_SOCIAL,
-                    NUMÉRO_FINESS_ÉTABLISSEMENT_MÉDICO_SOCIAL
+                    NUMÉRO_FINESS_ÉTABLISSEMENT_MÉDICO_SOCIAL,
                 ],
                 "annee": [2019, 2018, 2021, 2020, 2019, 2018],
                 "nombre_cdd_remplacement": [2.0, 19.0, 5.0, 5.0, 5.0, 5.0],
                 "taux_etp_vacants": [0.1197, 0.0483, 0.0, 0.0, 0.0, 0.0],
-                "taux_prestation_externes": [0.0232, NaN, 0.0164, 0.0164, 0.0082, 0.0164],
+                "taux_prestation_externes": [0.0232, nan, 0.0164, 0.0164, 0.0082, 0.0164],
                 "taux_rotation_personnel": [0.1923, 0.1429, 0.0352, 0.0352, 0.0141, 0.0352],
                 "taux_absenteisme_maladie_courte_duree": [0.0028, 0.0021, 0.0083, 0.0083, 0.0125, 0.0083],
                 "taux_absenteisme_maladie_moyenne_duree": [0.0465, 0.0717, 0.0166, 0.0166, 0.0149, 0.0166],
-                "taux_absenteisme_maladie_longue_duree": [0.0, .1194, 0.0089, 0.0089, 0.0319, 0.0089],
+                "taux_absenteisme_maladie_longue_duree": [0.0, 0.1194, 0.0089, 0.0089, 0.0319, 0.0089],
                 "taux_absenteisme_maternite_paternite": [0.0, 0.0, 0.0128, 0.0128, 0.0005, 0.0128],
-                "taux_absenteisme_accident_maladie_professionnelle": [0.0008, 0.0246, 0.0085, 0.0085, 0.0088,0.0085],
+                "taux_absenteisme_accident_maladie_professionnelle": [0.0008, 0.0246, 0.0085, 0.0085, 0.0088, 0.0085],
                 "taux_absenteisme_conges_speciaux": [0.0109, 0.0, 0.0004, 0.0004, 0.0, 0.0004],
                 "taux_absenteisme_hors_formation": [0.0609, 0.2179, 0.0554, 0.0554, 0.0685, 0.0554],
-                "nombre_etp_realises": [NaN, NaN, 188.5, 50.64999999999999, 17.19, 4.55],
+                "nombre_etp_realises": [nan, nan, 188.5, 50.64999999999999, 17.19, 4.55],
             }
         )
 
@@ -186,7 +195,14 @@ class TestAjouteLeBlocDesRessourcesHumainesMédicoSocial:
             base_de_données_test,
         )
 
-        pd.testing.assert_frame_equal(données_ressources_humaines_enregistrées.sort_index(axis=1), données_ressources_humaines_attendues.sort_index(axis=1))
+        pd.testing.assert_frame_equal(
+            données_ressources_humaines_enregistrées.sort_index(axis=1)
+            .sort_values(by=["numero_finess_etablissement_territorial", "annee"])
+            .reset_index(drop=True),
+            données_ressources_humaines_attendues.sort_index(axis=1)
+            .sort_values(by=["numero_finess_etablissement_territorial", "annee"])
+            .reset_index(drop=True),
+        )
 
     @patch.object(datacrawler, "sauvegarde")
     def test_revient_à_la_situation_initiale_si_l_écriture_des_indicateurs_ressources_humaines_échoue(self, mocked_sauvegarde: Mock) -> None:
@@ -228,17 +244,17 @@ class TestAjouteLeBlocDesRessourcesHumainesMédicoSocial:
 
         pd.testing.assert_frame_equal(table_des_ressources_humaines.sort_index(axis=1), table_des_ressources_humaines_existante.sort_index(axis=1))
 
-        date_du_fichier_ann_ca_ej_et = base_de_données_test.execute(
-            f"""SELECT * FROM {TABLE_DES_MISES_À_JOUR_DES_FICHIERS_SOURCES} WHERE fichier = '{FichierSource.DIAMANT_ANN_CA_EJ_ET.value}'"""
+        date_du_fichier_ann_ca_ej_et = base_de_données_test.connect().execute(
+            text(f"""SELECT * FROM {TABLE_DES_MISES_À_JOUR_DES_FICHIERS_SOURCES} WHERE fichier = '{FichierSource.DIAMANT_ANN_CA_EJ_ET.value}'""")
         )
         assert date_du_fichier_ann_ca_ej_et.fetchone() == (date(2020, 1, 1), FichierSource.DIAMANT_ANN_CA_EJ_ET.value)
 
-        date_du_fichier_ann_errd_ej_et = base_de_données_test.execute(
-            f"""SELECT * FROM {TABLE_DES_MISES_À_JOUR_DES_FICHIERS_SOURCES} WHERE fichier = '{FichierSource.DIAMANT_ANN_ERRD_EJ_ET.value}'"""
+        date_du_fichier_ann_errd_ej_et = base_de_données_test.connect().execute(
+            text(f"""SELECT * FROM {TABLE_DES_MISES_À_JOUR_DES_FICHIERS_SOURCES} WHERE fichier = '{FichierSource.DIAMANT_ANN_ERRD_EJ_ET.value}'""")
         )
         assert date_du_fichier_ann_errd_ej_et.fetchone() == (date(2020, 1, 2), FichierSource.DIAMANT_ANN_ERRD_EJ_ET.value)
 
-        date_du_fichier_ann_ms_tdp_et = base_de_données_test.execute(
-            f"""SELECT * FROM {TABLE_DES_MISES_À_JOUR_DES_FICHIERS_SOURCES} WHERE fichier = '{FichierSource.DIAMANT_ANN_MS_TDP_ET.value}'"""
+        date_du_fichier_ann_ms_tdp_et = base_de_données_test.connect().execute(
+            text(f"""SELECT * FROM {TABLE_DES_MISES_À_JOUR_DES_FICHIERS_SOURCES} WHERE fichier = '{FichierSource.DIAMANT_ANN_MS_TDP_ET.value}'""")
         )
         assert date_du_fichier_ann_ms_tdp_et.fetchone() == (date(2020, 1, 3), FichierSource.DIAMANT_ANN_MS_TDP_ET.value)